upgrade commons io from 1.4 to 2.2
